Created the scrollView with a for loop and a switch statement. Seemed a nicer way to go than having the whole thing being hard coded as in the test file.
Added global buttons for the currencies so that the selected currency can have it's colour changed.
These buttons will supply the names of the currencies that can be chosen from.
An action has been added to the button, but i don't think it will be used, as the action will come from the selection.