Needed to have a formatted string and also a number to calculate with (so we don't get an error because of the formatting) var textFieldOzzieDollarsFormatted_asString:String = String()
Added a shadow to UIView that is behind the horizontal scroller...
Unfortunately it's on top of the main layer and i cannot get the shadow to make it look like it's underneath - i should have looked at this before i designed the layout in my head.
Still looks ok.
	m.layer.shadowColor = UIColor.blackColor().CGColor
		m.layer.shadowOffset = CGSize(width: 0, height: 1)
		m.layer.shadowOpacity = 0.5
		m.layer.shadowRadius = 2

Draw a dotted line.
This was a pain, it seems very convoluted on how it works.
func drawCustomImage(size: CGSize) -> UIImage {
		UIGraphicsBeginImageContextWithOptions(CGSize(width: 500, height: 500), false, 0)
		let context = UIGraphicsGetCurrentContext()
		CGContextSetLineWidth(context, 30.0)
		CGContextMoveToPoint(context, 150, 55)
		CGContextAddLineToPoint(context, 400, 55)

		var dash:[CGFloat] = [0.0, 50.0]
		CGContextSetLineCap(context, kCGLineCapSquare)
		CGContextSetLineDash(context, 0.0, dash, 2)

		CGContextSetStrokeColorWithColor(context, UIColor.blackColor().CGColor)
		CGContextStrokePath(context)
		// Drawing complete, retrieve the finished image and cleanup
		let image = UIGraphicsGetImageFromCurrentImageContext()
		UIGraphicsEndImageContext()
		return image
	}

Needed to reset the text input, so i didn't get an error because of the formatting on a second run around.
func textFieldShouldBeginEditing(t: UITextField) -> Bool {
		self.textFieldOzzieDollars.text = ""
		return true
	}